Added a ProgressBar widget, essentially a non-interactive Slider
